requirements.txt
Babel==2.1.1
certifi==0.0.8
chardet==1.0.1
DateTime==4.0.1
decorator==4.0.4
Fabric==1.4.3
Flask==0.8
Flask-Babel==0.9
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
gunicorn==0.14.2
Jinja2==2.6
nltk==3.0.0
numpy==1.10.1
pbr==0.11.0
pycrypto==2.6.1
python-memcached==1.53
pytz==2015.7
requests==0.11.1
six==1.10.0
speaklater==1.3
SQLAlchemy==0.9.10
sqlalchemy-migrate==0.9.5
sqlparse==0.1.18
ssh==1.8.0
Tempita==0.5.2
Werkzeug==0.8.3
wheel==0.24.0
WTForms==2.0.2
zope.interface==4.1.3
